test: isolate single-pool transaction contracts
